# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BillForge API",
    "description": "BillForge - Modern Invoice Processing Platform API\n\n## Overview\n\nBillForge is a comprehensive accounts payable automation platform that helps businesses:\n- Capture and digitize invoices with OCR\n- Route invoices through customizable approval workflows\n- Manage vendor relationships\n- Generate financial reports and analytics\n\n## Authentication\n\nAll API endpoints (except /health) require authentication using JWT Bearer tokens.\n\n1. Obtain tokens via POST /api/v1/auth/login\n2. Include the access token in the Authorization header: `Authorization: Bearer <token>`\n3. Refresh tokens before expiry using POST /api/v1/auth/refresh\n\n## Multi-Tenancy\n\nBillForge is a multi-tenant system. Include tenant_id in authentication requests.\n",
    "contact": {
      "name": "BillForge Support",
      "email": "support@billforge.app"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/api/v1/bill-com/sync/vendors": {
      "post": {
        "tags": [
          "Bill.com"
        ],
        "summary": "Sync vendors from Bill.com",
        "operationId": "sync_vendors",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vendors synced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/netsuite/connect": {
      "post": {
        "tags": [
          "NetSuite"
        ],
        "summary": "Connect to NetSuite (save credentials & test connection)",
        "operationId": "netsuite_connect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetSuiteConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "NetSuite connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetSuiteConnectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials or auth not yet implemented"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/netsuite/disconnect": {
      "post": {
        "tags": [
          "NetSuite"
        ],
        "summary": "Disconnect NetSuite",
        "operationId": "netsuite_disconnect",
        "responses": {
          "200": {
            "description": "NetSuite disconnected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetSuiteDisconnectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/netsuite/status": {
      "get": {
        "tags": [
          "NetSuite"
        ],
        "summary": "Get NetSuite connection status",
        "operationId": "netsuite_status",
        "responses": {
          "200": {
            "description": "NetSuite status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetSuiteStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/netsuite/sync/vendors": {
      "post": {
        "tags": [
          "NetSuite"
        ],
        "summary": "Sync vendors from NetSuite",
        "operationId": "netsuite_sync_vendors",
        "responses": {
          "200": {
            "description": "Vendors synced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/quickbooks/sync/vendors": {
      "post": {
        "tags": [
          "QuickBooks"
        ],
        "summary": "Sync vendors from QuickBooks",
        "operationId": "sync_vendors",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncVendorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vendors synced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncVendorsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/sage-intacct/sync/vendors": {
      "post": {
        "tags": [
          "Sage Intacct"
        ],
        "summary": "Sync vendors from Sage Intacct",
        "operationId": "sync_vendors",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vendors synced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/vendors": {
      "get": {
        "tags": [
          "Vendors"
        ],
        "operationId": "list_vendors",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated vendor list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Vendors"
        ],
        "operationId": "create_vendor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vendor created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vendor"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NetSuiteConnectRequest": {
        "type": "object",
        "description": "NetSuite connection request",
        "required": [
          "account_id",
          "client_id",
          "client_secret"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "description": "NetSuite account ID (e.g. \"TSTDRV1234567\")"
          },
          "client_id": {
            "type": "string",
            "description": "OAuth 2.0 client ID"
          },
          "client_secret": {
            "type": "string",
            "description": "OAuth 2.0 client secret"
          }
        }
      },
      "NetSuiteConnectResponse": {
        "type": "object",
        "description": "Response body for a successful NetSuite connect.",
        "required": [
          "status",
          "account_id"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "description": "NetSuite account id that was registered."
          },
          "status": {
            "type": "string",
            "description": "Always \"connected\"."
          }
        }
      },
      "NetSuiteDisconnectResponse": {
        "type": "object",
        "description": "Response body for NetSuite disconnect.",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "Always \"disconnected\"."
          }
        }
      },
      "NetSuiteStatus": {
        "type": "object",
        "description": "NetSuite connection status",
        "required": [
          "connected",
          "sync_enabled"
        ],
        "properties": {
          "account_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "NetSuite account ID"
          },
          "connected": {
            "type": "boolean",
            "description": "Whether NetSuite is connected"
          },
          "last_sync_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Last sync timestamp"
          },
          "sync_enabled": {
            "type": "boolean",
            "description": "Sync enabled"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "description": "Pagination information",
        "required": [
          "page",
          "per_page",
          "total_items",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "Current page number (1-indexed)",
            "example": 1,
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "description": "Items per page",
            "example": 25,
            "minimum": 0
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of items (matches PaginationMeta.total_items on the wire)",
            "example": 100,
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "description": "Total number of pages",
            "example": 4,
            "minimum": 0
          }
        }
      },
...
      "SyncRequest": {
        "type": "object",
        "description": "Sync request",
        "required": [
          "full_sync"
        ],
        "properties": {
          "full_sync": {
            "type": "boolean",
            "description": "Force full sync (vs incremental)"
          }
        }
      },
      "SyncResponse": {
        "type": "object",
        "description": "Sync response",
        "required": [
          "imported",
          "updated",
          "skipped"
        ],
        "properties": {
          "imported": {
            "type": "integer",
            "format": "int64",
            "description": "Number of records imported",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "format": "int64",
            "description": "Number of records skipped",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "format": "int64",
            "description": "Number of records updated",
            "minimum": 0
          }
        }
      },
      "SyncVendorsRequest": {
        "type": "object",
        "description": "Sync vendors from QuickBooks",
        "required": [
          "full_sync"
        ],
        "properties": {
          "full_sync": {
            "type": "boolean",
            "description": "Force full sync (vs incremental)"
          }
        }
      },
      "SyncVendorsResponse": {
        "type": "object",
        "description": "Sync vendors response",
        "required": [
          "imported",
          "updated",
          "skipped",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "integer",
            "format": "int64",
            "description": "Number of vendors that failed to sync",
            "minimum": 0
          },
          "imported": {
            "type": "integer",
            "format": "int64",
            "description": "Number of vendors imported",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "format": "int64",
            "description": "Number of vendors skipped",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "format": "int64",
            "description": "Number of vendors updated",
            "minimum": 0
          }
        }
      },
...
      "Vendor": {
        "type": "object",
        "description": "Vendor record mirroring billforge_core::domain::Vendor",
        "required": [
          "id",
          "tenant_id",
          "name",
          "vendor_type",
          "status",
          "w9_on_file",
          "payment_hold",
          "contacts",
          "tags",
          "custom_fields",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "address": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VendorAddressInfo"
              }
            ]
          },
          "bank_account": {},
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VendorContactInfo"
            }
          },
          "created_at": {
            "type": "string"
          },
          "custom_fields": {},
          "default_department": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_gl_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_payment_method": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "legal_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string",
            "example": "Acme Corporation"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "payment_hold": {
            "type": "boolean"
          },
          "payment_hold_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "payment_terms": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "primary_contact": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VendorContactInfo"
              }
            ]
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tax_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "tax_id_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "tenant_id": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "vendor_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "vendor_type": {
            "type": "string"
          },
          "w9_on_file": {
            "type": "boolean"
          },
          "w9_received_date": {
            "type": [
              "string",
              "null"
            ]
          },
          "website": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "VendorAddressInfo": {
        "type": "object",
        "description": "Vendor address",
        "required": [
          "line1",
          "city",
          "postal_code",
          "country"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": [
              "string",
              "null"
            ]
          },
          "postal_code": {
            "type": "string"
          },
          "state": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "VendorContactInfo": {
        "type": "object",
        "description": "Vendor contact person",
        "required": [
          "id",
          "name",
          "is_primary"
        ],
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_primary": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "VendorList": {
        "type": "object",
        "description": "Paginated list of vendors",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vendor"
            },
            "description": "Vendor records"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo",
            "description": "Pagination information"
          }
        }
      },
...
    }
  },
...
}